pyproject.toml
[tool.poetry]
name = "bsystems-papss-backend"
version = "0.1.0"
description = ""
authors = ["Kojo <boakyeni@usc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-environ = "^0.11.2"
djoser = "^2.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
django-phonenumber-field = "^7.2.0"
phonenumbers = "^8.13.25"
mysqlclient = "^2.2.0"
django-countries = "^7.5.1"
drf-extra-fields = "^3.7.0"
django-mptt = "^0.15.0"
pillow = "^10.1.0"
celery = "^5.3.5"
django-celery-email = "^3.0.0"
flower = "^2.0.1"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
mysql-connector-python = "^8.2.0"
libmagic = "^1.0"
python-magic = "^0.4.27"
watchdog = "^3.0.0"
django-autoslug = "^1.9.9"
setuptools = "^68.2.2"
gunicorn = "^21.2.0"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
